requirements.txt
# Core astronomy libraries
astropy>=5.0
astroquery>=0.4.6

# Image processing
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0

# HTTP requests for API calls
requests>=2.25.0

# Optional: For better star detection (if needed)
# photutils>=1.5.0
# sep>=1.2.0

# Advanced constellation fitting
scipy>=1.10.0 